Move router into router/index.js
This is to match the new vue-cli defaults
