+++ src/packs/main.js
@@ -8,7 +8,7 @@ import '@/components/_globals.js';
 import '@/stylesheets/global.scss';
 import '@/stylesheets/tailwind.css';
 
-import router from '@/router';
+import router from '@/router/';
 import store from '@/store/index';
 
 Vue.config.productionTip = false;
